requirements.txt
streamlit
pandas
yfinance
python-dotenv
requests
streamlit-autorefresh